Renaming Ui Components folder to maintain pattern
